Unified UX for all evals
