+++ lynxius/evals/bert_score.py
@@ -1,5 +1,7 @@
 from lynxius.evals.evaluator import Evaluator
 from lynxius.rag.types import ContextChunk
+from lynxius_evals.evaluators.bert_score_eval import BertScoreEval
+from lynxius_evals.models.openai import OpenAIModel
 
 
 class BertScore(Evaluator):
@@ -10,6 +12,7 @@ class BertScore(Evaluator):
         presence_threshold: float = 0.65,
         href: str = None,
         tags: list[str] = [],
+        run_local: bool = False,
     ):
         levels = ["word", "sentence"]
         if level not in levels:
@@ -23,6 +26,8 @@ class BertScore(Evaluator):
         self.level = level
         self.presence_threshold = presence_threshold
         self.samples = []
+        self.run_local = run_local
+        self.evaluated_results = None
 
     def add_trace(self, reference: str, output: str, context: list[ContextChunk] = []):
         if not reference or not output:
@@ -31,23 +36,68 @@ class BertScore(Evaluator):
         self.samples.append((reference, output, context))
 
     def get_url(self):
-        return "/evals/run/bert_score/"
+        return (
+            "/evals/store/bert_score/" if self.run_local else "/evals/run/bert_score/"
+        )
 
     def get_request_body(self):
-        body = {
-            "label": self.label,
-            "href": self.href,
-            "tags": self.tags,
-            "level": self.level,
-            "presence_threshold": self.presence_threshold,
-            "data": [
-                {
-                    "reference": item[0],
-                    "output": item[1],
-                    "contexts": [c.__dict__ for c in item[2]],
-                }
-                for item in self.samples
-            ],
-        }
+        if self.run_local and self.evaluated_results is None:
+            raise Exception(
+                "Call evaluate_local() before storing the local evaluation output"
+            )
+
+        if self.run_local:
+            body = {
+                "label": self.label,
+                "href": self.href,
+                "tags": self.tags,
+                "level": self.level,
+                "presence_threshold": self.presence_threshold,
+                "version": "1",
+                "data": [
+                    {
+                        "reference": result["reference"],
+                        "output": result["output"],
+                        "precision": result["precision"],
+                        "recall": result["recall"],
+                        "f1": result["f1"],
+                        "missing_tokens": result["missing_tokens"],
+                        "contexts": [c.__dict__ for c in result["contexts"]],
+                    }
+                    for result in self.evaluated_results
+                ],
+            }
+        else:
+            body = {
+                "label": self.label,
+                "href": self.href,
+                "tags": self.tags,
+                "level": self.level,
+                "presence_threshold": self.presence_threshold,
+                "data": [
+                    {
+                        "reference": item[0],
+                        "output": item[1],
+                        "contexts": [c.__dict__ for c in item[2]],
+                    }
+                    for item in self.samples
+                ],
+            }
 
         return body
+
+    def evaluate_local(self):
+        model = OpenAIModel(embedding_model="text-embedding-3-small")
+        eval = BertScoreEval(model, self.level, self.presence_threshold)
+
+        variables = []
+        for sample in self.samples:
+            variables.append(
+                {
+                    "reference": sample[0],
+                    "output": sample[1],
+                    "contexts": sample[2],
+                }
+            )
+
+        self.evaluated_results = eval.evaluate(variables)
