+++ lynxius/evals/custom_eval.py
@@ -1,6 +1,10 @@
 from string import Formatter
+
 from lynxius.evals.evaluator import Evaluator
 from lynxius.rag.types import ContextChunk
+from lynxius_evals.evaluators.llm_based_eval import LLMBasedEval
+from lynxius_evals.models.openai import OpenAIModel
+from lynxius_evals.prompts.eval_prompt import EvalPromptTemplate
 
 
 class CustomEval(Evaluator):
@@ -11,6 +15,7 @@ class CustomEval(Evaluator):
         name: str = None,
         href: str = None,
         tags: list[str] = [],
+        run_local: bool = False,
     ):
         [Evaluator.validate_tag(value) for value in tags]
 
@@ -23,6 +28,8 @@ class CustomEval(Evaluator):
         self.variables = [
             fname for _, fname, _, _ in Formatter().parse(prompt_template) if fname
         ]
+        self.run_local = run_local
+        self.evaluated_results = None
 
     def add_trace(self, values: dict[str, str], context: list[ContextChunk] = []):
 
@@ -42,19 +49,65 @@ class CustomEval(Evaluator):
         self.samples.append((values, context))
 
     def get_url(self):
-        return "/evals/run/custom_eval/"
+        return (
+            "/evals/store/custom_eval/" if self.run_local else "/evals/run/custom_eval/"
+        )
 
     def get_request_body(self):
-        body = {
-            "label": self.label,
-            "href": self.href,
-            "tags": self.tags,
-            "prompt_template": self.prompt_template,
-            "name_override": self.name_override,
-            "data": [
-                {"variables": item[0], "contexts": [c.__dict__ for c in item[1]]}
-                for item in self.samples
-            ],
-        }
+        if self.run_local and self.evaluated_results is None:
+            raise Exception(
+                "Call evaluate_local() before storing the local evaluation output"
+            )
+
+        if self.run_local:
+            body = {
+                "label": self.label,
+                "href": self.href,
+                "tags": self.tags,
+                "prompt_template": self.prompt_template,
+                "name_override": self.name_override,
+                "version": "1",
+                "data": [
+                    {
+                        "llm_input": result["llm_input"],
+                        "llm_output": result["llm_output"],
+                        "variables": self.samples[i][0],
+                        "score": str(result["score"]),
+                        "contexts": [c.__dict__ for c in result["contexts"]],
+                    }
+                    for i, result in enumerate(self.evaluated_results)
+                ],
+            }
+        else:
+            body = {
+                "label": self.label,
+                "href": self.href,
+                "tags": self.tags,
+                "prompt_template": self.prompt_template,
+                "name_override": self.name_override,
+                "data": [
+                    {"variables": item[0], "contexts": [c.__dict__ for c in item[1]]}
+                    for item in self.samples
+                ],
+            }
 
         return body
+
+    def evaluate_local(self):
+        model = OpenAIModel()
+        template = EvalPromptTemplate("CustomEval", self.prompt_template)
+        eval = LLMBasedEval(
+            model,
+            template,
+            # TODO: now that we run this locally, we can expose output_map to be set
+            # manually.
+            output_map={"correct": True, "incorrect": False},
+            output_default=False,
+        )
+
+        variables = []
+        for sample in self.samples:
+            # Merge contexts and variables to be formatted in the prompt
+            variables.append(sample[0] | {"contexts": sample[1]})
+
+        self.evaluated_results = eval.evaluate(variables)
