--- lynxius/evals/local/evaluator_local.py
@@ -1,12 +1,0 @@
-from lynxius.evals.evaluator import Evaluator
-from abc import abstractmethod
-
-
-class EvaluatorLocal(Evaluator):
-    """
-    A base class for all Lynxius evals that are executed locally
-    """
-
-    @abstractmethod
-    def evaluate_local(self):
-        return NotImplemented
